pyproject.toml
[project]
name = "apkdownloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=1.27.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "typer>=0.15.2",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

line-length = 120
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
indent-style = "space"
indent-width = 4


[tool.pyright]
include = [
    "src",
    "tests",
]
exclude = [
    "**/__pycache__",
]
typeCheckingMode = "strict"

reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportAssertAlwaysTrue = true
reportMatchNotExhaustive = true
reportImplicitOverride = true
reportMissingSuperCall = true
reportUnusedCoroutine = true